+++ src/evalBlock.js
@@ -1,3 +1,29 @@
+const evalApi = require('./evalApi.js');
+
+
+// TODO run some basic checks against provided code,
+// like ensuring there are no uses of `Math.random`
+
+const evalFuncTemplate = `
+  const bml = this;
+  const __USER_DEFS = {};
+
+  function provide(obj) {
+    for (let key in obj) {
+      __USER_DEFS[key] = obj[key];
+    }
+  }
+
+  ////////// start user code
+
+  ***USER CODE SLOT***
+
+  ///////// end userspace code
+
+  return __USER_DEFS;
+`;
+
+
 class EvalBlock {
   constructor(string) {
     this.string = string;
@@ -5,6 +31,15 @@ class EvalBlock {
 
   toString() {
     return `EvalBlock('${this.string}')`;
+  }
+  
+  toFunc() {
+    let funcSrc = evalFuncTemplate.replace('***USER CODE SLOT***', this.string);
+    return new Function(funcSrc).bind(evalApi.api);
+  }
+  
+  execute() {
+    return this.toFunc()();
   }
 }